Rework herald

+++ app/helpers/application_helper.rb
@@ -1,14 +1,16 @@
 module ApplicationHelper
 
   def sharing_btn(user)
-    if user == current_user
-      link_to "That's you!", "#", :class => "btn btn-social", :rel => "tooltip", :data => { :title => "Yeah really..", :placement => "top" }
-    else
-      if current_user.isSharingWith(user)
-        friendship = current_user.friendships.where(:friend_id => user.id).first
-        link_to content_tag("i", "", :class => "icon-heart")+" Sharing", friendship_path(friendship), :method => :delete, :remote => true, :class => "btn btn-social btn-unsocial sharing-btn #{user.id}"
+    if @herald.nil?
+      if user == current_user
+        link_to "That's you!", "#", :class => "btn btn-social", :rel => "tooltip", :data => { :title => "Yeah really..", :placement => "top" }
       else
-        link_to "Share", friendships_path(:friend => user), :method => :post, :remote => true, :class => "btn btn-social sharing-btn #{user.id}"
+        if current_user.isSharingWith(user)
+          friendship = current_user.friendships.where(:friend_id => user.id).first
+          link_to content_tag("i", "", :class => "icon-heart")+" Sharing", friendship_path(friendship), :method => :delete, :remote => true, :class => "btn btn-social btn-unsocial sharing-btn #{user.id}"
+        else
+          link_to "Share", friendships_path(:friend => user), :method => :post, :remote => true, :class => "btn btn-social sharing-btn #{user.id}"
+        end
       end
     end
   end
